tutorials: Renamed for consistency + fixed tutorial code.
