+++ tutorials/whatsup/reference/main.go
@@ -39,7 +39,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus/collectors"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/prometheus/client_golang/tutorial/internal"
+	"github.com/prometheus/client_golang/tutorials/whatsup/internal"
 )
 
 func main() {
